fix: Fix unsupported OpenGL under macOS

+++ jsolex/src/main/java/me/champeau/a4j/jsolex/app/jfx/gl/Volume3DTexture.java
@@ -30,7 +30,7 @@ import static org.lwjgl.opengl.GL12.GL_TEXTURE_WRAP_R;
 import static org.lwjgl.opengl.GL12.glTexImage3D;
 import static org.lwjgl.opengl.GL13.GL_TEXTURE0;
 import static org.lwjgl.opengl.GL13.glActiveTexture;
-import static org.lwjgl.opengl.GL30.GL_R32F;
+import static org.lwjgl.opengl.GL30.GL_R16F;
 
 /**
  * Manages a 3D texture for volume rendering of spectral data.
@@ -91,7 +91,7 @@ public class Volume3DTexture {
         textureId = glGenTextures();
         glBindTexture(GL_TEXTURE_3D, textureId);
 
-        glTexImage3D(GL_TEXTURE_3D, 0, GL_R32F, width, height, depth, 0, GL_RED, GL_FLOAT, buffer);
+        glTexImage3D(GL_TEXTURE_3D, 0, GL_R16F, width, height, depth, 0, GL_RED, GL_FLOAT, buffer);
 
         glTexParameteri(GL_TEXTURE_3D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
         glTexParameteri(GL_TEXTURE_3D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
